+++ lib/src/helper.dart
@@ -105,6 +105,11 @@ class Helper {
   static Future<void> setSpeakerphoneOn(bool enable) =>
       NativeAudioManagement.setSpeakerphoneOn(enable);
 
+  /// Ensure audio session
+  /// for iOS only
+  static Future<void> ensureAudioSession() =>
+      NativeAudioManagement.ensureAudioSession();
+
   /// Enable speakerphone, but use bluetooth if audio output device available
   /// for iOS/Android only
   static Future<void> setSpeakerphoneOnButPreferBluetooth() =>