+++ lib/src/native/rtc_video_renderer_impl.dart
@@ -12,6 +12,7 @@ class RTCVideoRenderer extends ValueNotifier<RTCVideoValue>
     implements VideoRenderer {
   RTCVideoRenderer() : super(RTCVideoValue.empty);
   int? _textureId;
+  bool _disposed = false;
   MediaStream? _srcObject;
   StreamSubscription<dynamic>? _eventSubscription;
 
@@ -47,8 +48,10 @@ class RTCVideoRenderer extends ValueNotifier<RTCVideoValue>
 
   @override
   set srcObject(MediaStream? stream) {
+    if (_disposed) {
+      throw 'Can\'t set srcObject: The RTCVideoRenderer is disposed';
+    }
     if (textureId == null) throw 'Call initialize before setting the stream';
-
     _srcObject = stream;
     WebRTC.invokeMethod('videoRendererSetSrcObject', <String, dynamic>{
       'textureId': textureId,
@@ -58,39 +61,55 @@ class RTCVideoRenderer extends ValueNotifier<RTCVideoValue>
       value = (stream == null)
           ? RTCVideoValue.empty
           : value.copyWith(renderVideo: renderVideo);
-    });
+    }).catchError((e) {
+      print('Got exception for RTCVideoRenderer::setSrcObject: ${e.message}');
+    }, test: (e) => e is PlatformException);
   }
 
-  void setSrcObject({MediaStream? stream, String? trackId}) {
-    if (textureId == null) throw 'Call initialize before setting the stream';
-
+  Future<void> setSrcObject({MediaStream? stream, String? trackId}) async {
+    if (_disposed) {
+      throw 'Can\'t set srcObject: The RTCVideoRenderer is disposed';
+    }
+    if (_textureId == null) throw 'Call initialize before setting the stream';
     _srcObject = stream;
-    WebRTC.invokeMethod('videoRendererSetSrcObject', <String, dynamic>{
-      'textureId': textureId,
-      'streamId': stream?.id ?? '',
-      'ownerTag': stream?.ownerTag ?? '',
-      'trackId': trackId ?? '0'
-    }).then((_) {
+    var oldTextureId = _textureId;
+    try {
+      await WebRTC.invokeMethod('videoRendererSetSrcObject', <String, dynamic>{
+        'textureId': _textureId,
+        'streamId': stream?.id ?? '',
+        'ownerTag': stream?.ownerTag ?? '',
+        'trackId': trackId ?? '0'
+      });
       value = (stream == null)
           ? RTCVideoValue.empty
           : value.copyWith(renderVideo: renderVideo);
-    });
+    } on PlatformException catch (e) {
+      throw 'Got exception for RTCVideoRenderer::setSrcObject: textureId $oldTextureId [disposed: $_disposed] with stream ${stream?.id}, error: ${e.message}';
+    }
   }
 
   @override
   Future<void> dispose() async {
+    if (_disposed) return;
     await _eventSubscription?.cancel();
     _eventSubscription = null;
     if (_textureId != null) {
-      await WebRTC.invokeMethod('videoRendererDispose', <String, dynamic>{
-        'textureId': _textureId,
-      });
-      _textureId = null;
+      try {
+        await WebRTC.invokeMethod('videoRendererDispose', <String, dynamic>{
+          'textureId': _textureId,
+        });
+        _textureId = null;
+        _disposed = true;
+      } on PlatformException catch (e) {
+        throw 'Failed to RTCVideoRenderer::dispose: ${e.message}';
+      }
     }
+
     return super.dispose();
   }
 
   void eventListener(dynamic event) {
+    if (_disposed) return;
     final Map<dynamic, dynamic> map = event;
     switch (map['event']) {
       case 'didTextureChangeRotation':
@@ -126,6 +145,9 @@ class RTCVideoRenderer extends ValueNotifier<RTCVideoValue>
 
   @override
   set muted(bool mute) {
+    if (_disposed) {
+      throw Exception('Can\'t be muted: The RTCVideoRenderer is disposed');
+    }
     if (_srcObject == null) {
       throw Exception('Can\'t be muted: The MediaStream is null');
     }